some code cleanup for db listeners, add typing to responses
